+++ experimental/packages/sdk-logs/src/types.ts
@@ -106,11 +106,24 @@ export interface LogRecordLimits {
   attributeCountLimit?: number;
 }
 
-export interface SimpleLogRecordProcessorOptions {
+/**
+ * Common options for SDK log processors.
+ */
+export interface LogRecordProcessorOptions {
+  /**
+   * A meter provider to which to record self-observability log processor metrics.
+   * @experimental This option is experimental and is subject to breaking changes in minor releases.
+   */
+  selfObsMeterProvider?: MeterProvider;
+}
+
+export interface SimpleLogRecordProcessorOptions
+  extends LogRecordProcessorOptions {
   exporter: LogRecordExporter;
 }
 
-export interface BatchLogRecordProcessorOptions {
+export interface BatchLogRecordProcessorOptions
+  extends LogRecordProcessorOptions {
   exporter: LogRecordExporter;
 
   /** The maximum batch size of every export. It must be smaller or equal to